fixing unit tests
