+++ src/ui/public/vis/__tests__/_vis.js
@@ -24,8 +24,6 @@ import expect from 'expect.js';
 import { VisProvider } from '..';
 import FixturesStubbedLogstashIndexPatternProvider from 'fixtures/stubbed_logstash_index_pattern';
 import { VisTypesRegistryProvider } from '../../registry/vis_types';
-import { DataAdapter, RequestAdapter } from '../../inspector/adapters';
-import { Inspector } from '../../inspector/inspector';
 
 describe('Vis Class', function () {
   let indexPattern;
@@ -116,145 +114,6 @@ describe('Vis Class', function () {
       const vis = new Vis(indexPattern);
       expect(vis.isHierarchical()).to.be(false);
     });
-  });
-
-  describe('inspector', () => {
-
-    describe('hasInspector()', () => {
-      it('should forward to inspectors hasInspector', () => {
-        const vis = new Vis(indexPattern, state({
-          inspectorAdapters: {
-            data: true,
-            requests: true,
-          }
-        }));
-        sinon.spy(Inspector, 'isAvailable');
-        vis.hasInspector();
-        expect(Inspector.isAvailable.calledOnce).to.be(true);
-        const adapters = Inspector.isAvailable.lastCall.args[0];
-        expect(adapters.data).to.be.a(DataAdapter);
-        expect(adapters.requests).to.be.a(RequestAdapter);
-      });
-
-      it('should return hasInspectors result', () => {
-        const vis = new Vis(indexPattern, state({}));
-        const stub = sinon.stub(Inspector, 'isAvailable');
-        stub.returns(true);
-        expect(vis.hasInspector()).to.be(true);
-        stub.returns(false);
-        expect(vis.hasInspector()).to.be(false);
-      });
-
-      afterEach(() => {
-        Inspector.isAvailable.restore();
-      });
-    });
-
-    describe('openInspector()', () => {
-
-      beforeEach(() => {
-        sinon.stub(Inspector, 'open');
-      });
-
-      it('should call openInspector with all attached inspectors', () => {
-        const Foodapter = class {};
-        const vis = new Vis(indexPattern, state({
-          inspectorAdapters: {
-            data: true,
-            custom: {
-              foo: Foodapter
-            }
-          }
-        }));
-        vis.openInspector();
-        expect(Inspector.open.calledOnce).to.be(true);
-        const adapters = Inspector.open.lastCall.args[0];
-        expect(adapters).to.be(vis.API.inspectorAdapters);
-      });
-
-      it('should pass the vis title to the openInspector call', () => {
-        const vis = new Vis(indexPattern, { ...state(), title: 'beautifulVis' });
-        vis.openInspector();
-        expect(Inspector.open.calledOnce).to.be(true);
-        const params = Inspector.open.lastCall.args[1];
-        expect(params.title).to.be('beautifulVis');
-      });
-
-      afterEach(() => {
-        Inspector.open.restore();
-      });
-    });
-
-    describe('inspectorAdapters', () => {
-
-      it('should register none for none requestHandler', () => {
-        const vis = new Vis(indexPattern, state({ requestHandler: 'none' }));
-        expect(vis.API.inspectorAdapters).to.eql({});
-      });
-
-      it('should attach data and request handler for courier', () => {
-        const vis = new Vis(indexPattern, state({ requestHandler: 'courier' }));
-        expect(vis.API.inspectorAdapters.data).to.be.a(DataAdapter);
-        expect(vis.API.inspectorAdapters.requests).to.be.a(RequestAdapter);
-      });
-
-      it('should allow enabling data adapter manually', () => {
-        const vis = new Vis(indexPattern, state({
-          requestHandler: 'none',
-          inspectorAdapters: {
-            data: true,
-          }
-        }));
-        expect(vis.API.inspectorAdapters.data).to.be.a(DataAdapter);
-      });
-
-      it('should allow enabling requests adapter manually', () => {
-        const vis = new Vis(indexPattern, state({
-          requestHandler: 'none',
-          inspectorAdapters: {
-            requests: true,
-          }
-        }));
-        expect(vis.API.inspectorAdapters.requests).to.be.a(RequestAdapter);
-      });
-
-      it('should allow adding custom inspector adapters via the custom key', () => {
-        const Foodapter =  class {};
-        const Bardapter = class {};
-        const vis = new Vis(indexPattern, state({
-          requestHandler: 'none',
-          inspectorAdapters: {
-            custom: {
-              foo: Foodapter,
-              bar: Bardapter,
-            }
-          }
-        }));
-        expect(vis.API.inspectorAdapters.foo).to.be.a(Foodapter);
-        expect(vis.API.inspectorAdapters.bar).to.be.a(Bardapter);
-      });
-
-      it('should not share adapter instances between vis instances', () => {
-        const Foodapter = class {};
-        const visState = state({
-          inspectorAdapters: {
-            data: true,
-            custom: {
-              foo: Foodapter
-            }
-          }
-        });
-        const vis1 = new Vis(indexPattern, visState);
-        const vis2 = new Vis(indexPattern, visState);
-        expect(vis1.API.inspectorAdapters.foo).to.be.a(Foodapter);
-        expect(vis2.API.inspectorAdapters.foo).to.be.a(Foodapter);
-        expect(vis1.API.inspectorAdapters.foo).not.to.be(vis2.API.inspectorAdapters.foo);
-        expect(vis1.API.inspectorAdapters.data).to.be.a(DataAdapter);
-        expect(vis2.API.inspectorAdapters.data).to.be.a(DataAdapter);
-        expect(vis1.API.inspectorAdapters.data).not.to.be(vis2.API.inspectorAdapters.data);
-      });
-    });
-
   });
 
   describe('vis addFilter method', () => {
